+++ R/summary.R
@@ -116,8 +116,13 @@ print.marginaleffects.summary <- function(x,
 #' `marginalmeans` function
 #' @family summary
 #' @export
-summary.marginalmeans <- function(object, conf_level = 0.95, ...) {
-    out <- tidy(object, conf_level = conf_level, ...)
+summary.marginalmeans <- function(
+    object,
+    transform_avg = NULL,
+    conf_level = 0.95,
+    ...) {
+
+    out <- tidy(object, conf_level = conf_level, transform_avg = transform_avg, ...)
     class(out) <- c("marginalmeans.summary", class(out))
     attr(out, "type") <- attr(object, "type")
     attr(out, "model_type") <- attr(object, "model_type")
@@ -210,8 +215,13 @@ print.marginalmeans.summary <- function(x,
 #' `predictions` function
 #' @family summary
 #' @export
-summary.predictions <- function(object, ...) {
-    out <- tidy(object, ...)
+summary.predictions <- function(
+    object, 
+    conf_level = 0.95,
+    transform_avg = NULL,
+    ...) {
+
+    out <- tidy(object, conf_level = conf_level, transform_avg = transform_avg, ...)
     class(out) <- c("predictions.summary", class(out))
     attr(out, "type") <- attr(object, "type")
     attr(out, "model_type") <- attr(object, "model_type")