+++ inst/tinytest/test-predictions.R
@@ -26,6 +26,33 @@ expect_error(predictions(mod, variables = list(2)), pattern = "names")
 expect_error(predictions(mod, variables = "am"), pattern = "list")
 
 
+# average prediction with delta method are asymptotically equivalent to back transformed
+set.seed(1024)
+N <- 1e5
+dat <- data.frame(
+    y = rbinom(N, 1, prob = .9),
+    x = rnorm(N))
+mod <- glm(y ~ x, family = binomial, data = dat)
+p1 <- tidy(predictions(mod)) # average prediction outside [0,1]
+p2 <- tidy(predictions(mod, type = "link"), transform_avg = insight::link_inverse(mod)) # average prediction inside [0,1]
+expect_equivalent(p1$estimate, p2$estimate, tolerance = .001)
+expect_equivalent(p1$conf.low, p2$conf.low, tolerance = .01)
+expect_equivalent(p1$conf.high, p2$conf.high, tolerance = .01)
+
+# average prediction with delta method can produce CI outside [0,1] interval
+set.seed(1022)
+N <- 10
+dat <- data.frame(
+    y = rbinom(N, 1, prob = .8),
+    x = rnorm(N))
+mod <- glm(y ~ x, family = binomial, data = dat)
+p1 <- tidy(predictions(mod)) # average prediction outside [0,1]
+p2 <- tidy(predictions(mod, type = "link"), transform_avg = insight::link_inverse(mod)) # average prediction outside [0,1]
+expect_true(p1$conf.high > 1)
+expect_true(p2$conf.high < 1)
+
+
+
 ################
 #  conf.level  #
 ################
@@ -43,7 +70,7 @@ for (L in c(.4, .7, .9, .95, .99, .999)) {
 #################################
 #  average adjusted predictions #
 #################################
-dat <- mtcars
+dadat <- mtcars
 dat$w <- 1:32
 mod <- lm(mpg ~ hp + am, dat)
 pre1 <- predictions(mod, by = "am")