+++ NEWS.md
@@ -8,6 +8,7 @@
 * `hypothesis`: The matrix column names are used as labels for hypothesis tests.
 * `hypothesis`: Better labels with "sequential", "reference", "pairwise".
 * `wts` argument is respected in `by` argument and with `*avg` shortcuts in the `transform_pre` argument.
+* `tidy.predictions()` and `tidy.marginalmeans()` get a new `transform_avg` argument.
 * New vignette: Unit-level contrasts in logistic regressions. Thanks to @arthur-albuquerque.
 * New vignette: Bootstrap example in standard errors vignette.